moved test_target_device to the common test scope
